+++ app/Navigation.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+import App from './index'; // Your main app component
+
+const Stack = createStackNavigator();
+
+export default function Navigation() {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator screenOptions={{ headerShown: false }}>
+        <Stack.Screen name="Home" component={App} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
