+++ app/main.js
@@ -1,0 +1,4 @@
+import { registerRootComponent } from 'expo';
+import Navigation from './Navigation'; // Import Navigation setup
+
+registerRootComponent(Navigation); // Register Navigation as the root component
