+++ react-native.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    assets: ['./assets/fonts'],  // Pointing to the fonts folder
+  };
+
